Move out vg paths
